# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RMA Account Statement API",
    "description": "Read-only access to imported account statement transactions.",
    "version": "1.0.0"
  },
  "paths": {
    "/accounts/{account_number}/transactions": {
      "get": {
        "summary": "Read Account Transactions",
        "operationId": "read_account_transactions_accounts__account_number__transactions_get",
        "parameters": [
          {
            "name": "account_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Number"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start of the date range (inclusive), e.g. 2026-09-01",
              "title": "From Date"
            },
            "description": "Start of the date range (inclusive), e.g. 2026-09-01"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End of the date range (inclusive), e.g. 2026-09-30",
              "title": "To Date"
            },
            "description": "End of the date range (inclusive), e.g. 2026-09-30"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionOut"
                  },
                  "title": "Response Read Account Transactions Accounts  Account Number  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_number}/cards": {
      "get": {
        "summary": "Read Account Cards",
        "operationId": "read_account_cards_accounts__account_number__cards_get",
        "parameters": [
          {
            "name": "account_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardOut"
                  },
                  "title": "Response Read Account Cards Accounts  Account Number  Cards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}": {
      "get": {
        "summary": "Read Card",
        "operationId": "read_card_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/card-products": {
      "get": {
        "summary": "Read Card Products",
        "operationId": "read_card_products_card_products_get",
        "parameters": [
          {
            "name": "card_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by 'debit' or 'credit'",
              "title": "Card Type"
            },
            "description": "Filter by 'debit' or 'credit'"
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardProductOut"
                  },
                  "title": "Response Read Card Products Card Products Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/card-products/{card_product_id}": {
      "get": {
        "summary": "Read Card Product",
        "operationId": "read_card_product_card_products__card_product_id__get",
        "parameters": [
          {
            "name": "card_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Product Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}/transactions": {
      "get": {
        "summary": "Read Card Transactions",
        "operationId": "read_card_transactions_cards__card_id__transactions_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start of the date range (inclusive)",
              "title": "From Date"
            },
            "description": "Start of the date range (inclusive)"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End of the date range (inclusive)",
              "title": "To Date"
            },
            "description": "End of the date range (inclusive)"
          },
          {
            "name": "amount_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum amount (foreign-currency or INR)",
              "title": "Amount Min"
            },
            "description": "Minimum amount (foreign-currency or INR)"
          },
          {
            "name": "amount_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum amount (foreign-currency or INR)",
              "title": "Amount Max"
            },
            "description": "Maximum amount (foreign-currency or INR)"
          },
          {
            "name": "merchant_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text merchant search, e.g. 'Wisdom Property'",
              "title": "Merchant Text"
            },
            "description": "Free-text merchant search, e.g. 'Wisdom Property'"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionOut"
                  },
                  "title": "Response Read Card Transactions Cards  Card Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}/forex-summary": {
      "get": {
        "summary": "Read Card Forex Summary",
        "operationId": "read_card_forex_summary_cards__card_id__forex_summary_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "as_of_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end date; defaults to today",
              "title": "As Of Date"
            },
            "description": "Window end date; defaults to today"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForexSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}/category-breakdown": {
      "get": {
        "summary": "Read Card Category Breakdown",
        "operationId": "read_card_category_breakdown_cards__card_id__category_breakdown_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start of the date range (inclusive)",
              "title": "From Date"
            },
            "description": "Start of the date range (inclusive)"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End of the date range (inclusive)",
              "title": "To Date"
            },
            "description": "End of the date range (inclusive)"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategorySpendOut"
                  },
                  "title": "Response Read Card Category Breakdown Cards  Card Id  Category Breakdown Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card_id}/recommendation": {
      "get": {
        "summary": "Read Card Recommendation",
        "operationId": "read_card_recommendation_cards__card_id__recommendation_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardRecommendationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/card-applications": {
      "post": {
        "summary": "Submit Card Application",
        "operationId": "submit_card_application_card_applications_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "account_number": {
            "type": "string",
            "title": "Account Number"
          },
          "card_product_id": {
            "type": "integer",
            "title": "Card Product Id"
          },
          "last4": {
            "type": "string",
            "title": "Last4"
          },
          "network": {
            "type": "string",
            "title": "Network"
          },
          "card_type": {
            "type": "string",
            "title": "Card Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_number",
          "card_product_id",
          "last4",
          "network",
          "card_type",
          "status",
          "issued_at",
          "created_at"
        ],
        "title": "CardOut"
      },
...
      "CategorySpendOut": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          }
        },
        "type": "object",
        "required": [
          "category",
          "total_amount",
          "transaction_count"
        ],
        "title": "CategorySpendOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}